Introduce `ct_suite_variant` macro
ct_suite_variant works like ct_suite, but without declaring the
compilation of the *_SUITE.erl files. This allows for what are in
effect multiple calls to ct_suite. The same result could be achieved
with the use of the 'matrix' paramter to `ct_suite`, though the use of
`ct_suite_variant` may be more readable depending on the context.