Add a second style of test sharding to ct_sharded_test
The group mode assumes that one shard per group is desired, and if the
shard_count specified in bazel does not match, the test will simply
fail. Since the only sharding modes bazel supports are 'explicit' and
'disabled', this seems reasonable. This mode aims to strike a better
balance with manual sharding where cases within a group are not spread
across shards, but unlike manual sharding is not prone to skipping
tests by accident.

The original sharding mode is now called 'uniform' sharding.