The dialyze rule now only analyzes the target and not its deps
The dialyze rule no longer analyzes the deps of the target. This was
probably not desired to begin with but when unnoticed before the
introduction of the warnings_as_errors attribute. A new attribute has
been added to the plt rule, deps, so that information for the dialyzer
about those deps can be fed to the dialyze rule. The transitive deps
are now included in that process as well.