Add support for stamping in .app files
If the --stamp flag is given as well as a --workspace_status_command,
that produces a STABLE_GIT_COMMIT value, that value is injected as the
vsn in .app files.

Since the same workspace command applies for all erlang apps in the
workspace, stamping is turned off for all deps.