Allow sending the erlang_bytecode rules to a macro
This keeps the BUILD file for larger erlang apps much more
succint. Controlled by a directive, and off by default.

Ran into some rough edges on the rule.LoadMacroFile function, as using
the native.filegroup within the macro would cause it to be duplicated
on re-runs. Worked around it by defining a rules_erlang macro that is
equivalent and can be called without the native. prefix.