erlang_bytecode2 compiles with transitive deps
since gazelle generates minimal deps, we need to include the
transitive deps so that, for instance, headers include headers of deps
are present.

For compilation, we actually don't need all the files from a dep, just
public headers and compiled behaviours. In fact, only the transient
headers are needed, not the transient behaviours. So this could be
optimized.