Track gazelle:excludes for resolving umbrella apps correctly
Previously it was assumed that the Configure pass walked the whole
before generation, but this was not true. It meant the approach of
using the presence of Config to know a directory was not excluded was
not sufficient. Now we track the 'exclude' directive ourselves to know
if we should skip a directory when resolving erlang apps within the
umbrella repo in gazelle