Set ignore_warnings true by default in the plt rule
The plt is used to cache type info for use in the dialyze rule. It's
unusual to care about these warnings, as they don't actually apply to
the target of the dialyze rule, and can be caused simply by the lack
of a complete graph of applications passed to the plt rule.