Add the optional_applications key to .app files
For now this is not intergrated with the erlang_app macro, as it's not
clear what the ideal usage would be

Should that macro have a new optional_deps key mirroring erlang.mk?
Should it just expose the arg from the underlying app_file rule?
In any case, that can come as a second PR