New heuristic for priv file relative paths with extract_many rule
this is still rather fragile. The only non-fragile solution that I can
think of currently is to copy the priv files into the outs (and
license files too) in the compile_many rule. This is not very
efficient, but otherwise since priv files may or not be generated,
their roots are not normalized. It might be possible to use the
erlang_app_sources macro to embed the paths as strings, and have the
maps in the provider, but it seems like a lot of complexity...