Always set exec_properties for compile_many macro
while these are buildbuddy specific, they aren't harmful for local
execution

The proper way to do this with exec_compatible_with and contraint,
that will allow selecting an exec platform on RBE with these
properties. However, we would want things to still work by default, so
that means we would have to always register a platform, that is empty
otherwise, that supports it.